1 giant unwanted (COMPLETELY UNWANTED) commit :<
I was successfully committing files individually. Then I lost track of
which ones I committed for to be updated on github. So I threw it all
together. And I tried to open it via the URL, and it resorted to my OLD
/last commit.

Note to Alan: I hope this doesn't give me -infinity :< it took a whole
semester to learn Github but I think I got it now for next project :D.
If you look at PCB, maybe it will cheer you up. Sorry